+++ src/test/java/com/ll/exam/sbb/repository/QuestionRepositoryTest.java
@@ -21,7 +21,7 @@ public class QuestionRepositoryTest {
 
     @Autowired
     private QuestionRepository questionRepository;
-    private static int lastSampleDataId;
+    private static long lastSampleDataId;
 
     @BeforeEach
     void beforeEach() {
@@ -34,7 +34,7 @@ public class QuestionRepositoryTest {
         questionRepository.truncateTable();
     }
 
-    public static int createSampleData(QuestionRepository questionRepository) {
+    public static long createSampleData(QuestionRepository questionRepository) {
         Question q1 = new Question();
         q1.setSubject("sbb가 무엇인가요?");
         q1.setContent("sbb에 대해서 알고 싶습니다.");
@@ -80,7 +80,7 @@ public class QuestionRepositoryTest {
     void 삭제() {
         assertThat(questionRepository.count()).isEqualTo(lastSampleDataId);
 
-        Question q = this.questionRepository.findById(1).get();
+        Question q = this.questionRepository.findById(1L).get();
         questionRepository.delete(q);
 
         assertThat(questionRepository.count()).isEqualTo(lastSampleDataId - 1);
@@ -88,11 +88,11 @@ public class QuestionRepositoryTest {
 
     @Test
     void 수정() {
-        Question q = this.questionRepository.findById(1).get();
+        Question q = this.questionRepository.findById(1L).get();
         q.setSubject("수정된 제목");
         questionRepository.save(q);
 
-        q = this.questionRepository.findById(1).get();
+        q = this.questionRepository.findById(1L).get();
 
         assertThat(q.getSubject()).isEqualTo("수정된 제목");
     }
@@ -146,7 +146,7 @@ public class QuestionRepositoryTest {
     @Test
     void findAAllPageable() {
         // Pageble : 한 페이지에 몇 개의 아이템이 나와야 하는지 + 현재 몇 페이지인지
-        PageRequest pageRequest = PageRequest.of(0, lastSampleDataId);
+        PageRequest pageRequest = PageRequest.of(0, (int)lastSampleDataId);
         Page<Question> page = questionRepository.findAll(pageRequest);
 
         assertThat(page.getTotalPages()).isEqualTo(1);
